Moved `PeriodInterface` to `DateTime` namespace
