+++ test/functional/DateTime/PeriodInterfaceTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace RebelCode\Diary\Test;
+namespace RebelCode\Diary\FuncTest\DateTime;
 
 use Xpmock\TestCase;
 
 /**
- * Tests {@see \RebelCode\Diary\PeriodInterface}.
+ * Tests {@see \RebelCode\Diary\DateTime\PeriodInterface}.
  *
  * @since 0.1
  */
@@ -14,14 +14,14 @@ class PeriodInterfaceTest extends TestCase
     /**
      * The name of the test subject.
      */
-    const TEST_SUBJECT_CLASSNAME = '\\RebelCode\\Diary\\PeriodInterface';
+    const TEST_SUBJECT_CLASSNAME = '\\RebelCode\\Diary\\DateTime\\PeriodInterface';
 
     /**
      * Creates a new instance of the test subject.
      *
      * @since 0.1
      *
-     * @return \RebelCode\Diary\PeriodInterface
+     * @return \RebelCode\Diary\DateTime\PeriodInterface
      */
     public function createInstance()
     {
@@ -37,7 +37,7 @@ class PeriodInterfaceTest extends TestCase
     /**
      * Tests whether a valid instance of the test subject can be created.
      *
-     * @covers \RebelCode\Diary\PeriodInterface
+     * @covers \RebelCode\Diary\DateTime\PeriodInterface
      *
      * @since 0.1
      */
